Assignmetns/Assessment2/requirements.txt
langgraph 
langchain-openai 
langchain_community
langchain 
pymilvus 
psycopg2-binary 
nltk 
scikit-learn 
pandas 
python-dotenv
PyPDF2
torch
transformers
chromadb 
sentence-transformers 
numpy
pypdf
langchain-milvus
pymilvus[model]